requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# LangChain ecosystem
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-mistralai>=0.2.0
langchain-chroma>=0.1.0

# New tech stack
langgraph>=0.2.0
chainlit==1.3.2
chromadb>=0.4.0
pyvis>=0.3.2

# Utilities
markdown2>=2.4.0
xhtml2pdf>=0.2.11
networkx>=3.0